Collaboration tools (#1337)
* Adds initial mockup for Chat interface.

* Adds models and makes debugging code easier to distinguish

* Transitions XMPP connections to WebSockets and removes delay when sending
initial messages.

* Adds new image picker component for selecting players by their image.

* Adds new chat functionality

* Breaks Chat management into external service.

* Fixes issue with unsubscribing from messages and sends UI Notifications

* Notifications are only sent if chat is not the foreground.

- Delayed Private Messages do not sent notifications.

* Fixes issues with authentication in and out.

- Validatation now actually sends a correct Oauth request.
- Logout logs out of both OAuth and Session.:

* Adds auto hide tabs when not connected to a party.

* Adds room rosters to the Chat Service

* Fixes issue with room connection subscriptions

* Fixes issues with authentication, moves logic for resub-ing to service, and
adds chat to player

* Fixes issue with tabs

* Dashboard hook up (#1310)

* Filled out rest of character card fields.

* pCard fixes.

* Started making player card model that converts pCards into a useable format for the view.

* Finishe setting up model and view.

* Many fixes to everything.

* Player card fully completed. Added ability to delete items.

* A player deletes all their cards before publishing one.

* Refines who listens to notifications and what notifications are sent.

* Proficiency now updates as it should. Need to fix passive perception/int updating.

* Minor fixes and documentation.

* Adds pCards in card service. Splits up skills notification.

* Linter fixes

* Fixes misnamed file

* Changes wording for party modal.
Adds playerName to card.

* Sets default gravatar image.

* Added delete button for saved parties.

* Update Readme

* Removes un-used svgs.

* Adds delete for party list in party manager UI.

* Converts party notion to group chat from pubsub node

* Fixes merge issue

* Changes publishing stuff and subscribing to people.

* Multiple chat fixes

- Party chats can no longer be deleted in the UI. Only leaving the chat will
  delete the overall party chat.
- The logic for creating a chat and managing it has been fully moved to the
  ChatService leaving the PartyManagerVM as a dumb listener (as it should be).
- Group chats now work again.
- Party chats should work too.

* Fixes joining previous chats.

* Adds auto-reconnection to XMPP Connection.

* Fixes merge issue that broke room connections.

* Adds new presence detection features

- App can now tell when another user has joined/left not just the user.
- Fixed an issue with presence subscriptions not being answered

* Fixes issue with presence subscriptions getting sent to the wrong jids

* Removes old plugin

* Removes old include

* Fixes issues with inviting users to chat, and other UI improvements

* Fixes issue where presence subscriptions are not accepted.

Solution: Full JIDs are not accepted.

* Adds party status line.

* Removes debug logging

* Adds new character picker before app load. Also changes look of switching modal

* Fixes some issues with cards updating.

* Adds system messages to chat.

* Fixes issue with Auth Service causing websockets to break

* FIxes issue with party manager not loading

* Changes to party manager modal.

* Players are now removed from the party when they leave the room.

* Fixes issues with login/switching and native XMPP auth.

- Calls to the publish method refuse to send if the connection is not
  established.

* Fixes character switching issues when connected to a room.

- User JIDs no longer contain the character ID as the resource in favor of a
  static "Adventurers Codex (Web) resource. The dynamic resource was causing
issues with room management and other XMPP features.

- When switching characters, the user is removed from all rooms they're a
  part of.

- Node JID URL fragment is now called party_node and only applies to the first
  character selected. Subsequent characters will not be "auto-joined" to the
given room. This is for feature simplicity.

* Character switching improvements.

- Attempting to switch characters to the current active character no longer
actually does the switch. It was too costly to perform the switch for no reason
and resulted in unexpected behavior.

* Adds text to party link/id in manager modal.

* Upon successful presence subscription, a request to subscribe to the users pubsub node will be sent.

* When a user leaves a party, they are unsubscribed from. This prevents their cards from being pulled when they aren't in the same party.

* Starts implementation of dropbox saver.

* Attempt to fix getting cards on joining a party

* Adds filter to only display cards of people that are in the room. Also, clears pCards when one leaves the party.

* Party manager modal fixes.

* Update chat_message.js

* Multiple chat/card service fixes.

- Adds preliminary methods to show chat bubbles in chat UI.
- Moves the players.changed notification to the CardService so that it fires
  for all player types.
- players.changed now sends pCards instead of PlayerCard models.
- Removes init-character JID overrides used in testing.

* Adds chat bubble ui to detail view and chat list.

- Fixes issue where notification failure would cause incorrect rosters to be
  saved

* Adds chat bubbles to msg log

* Fixes issue with chat sorting and adds character name to chat line

* Adds check for non-existant cards

* Party fixes

- Adds DM Card publishing
- Adds exception for dm party list for the DM.

* Adds new check for invalid card

* Minor wiring fixes in creating character cards.
Refactors subscribing to a node and fetching items into re-usable methods.

* Inspiration status line now updates with inspiration.
Fixed issue with proficiency not updating based on level change.
Added hotkeys and made chat and party ones only trigger when you're in a party.

* Linter fixes.

* Adds "Save to Notes" to chat messages. (#1333)

* Implements "Save to Chat" feature.

- All chat messages have a button next to them which "saves them to the notes".
    - This creates a new note called "Saved from Chat" if one like it doesn't
      already exist and saves the message to it.

* Save to Notes no longer appears for DMs.

- DMs don't have multiple notes so creating one doesnt make any sence atm.
- Removes debugging code.

* Removes debugging logic

* Updates Auth flow (#1334)

- Invalid tokens in the URL no longer cause login to display empty name field.
- Adds automatic retry with local token if URL token is invalid (bookmarked links).

* Fixes character import/export unicode parsing issue (#1335)

* Fix linter issues. Comment out failing test.

* Made status line more robust and accurate.
Hotkeys don't trigger when editing character/dm name.
Removed actions well hotkey.

* Init fixes

* Adds more logging to XMPP Service errors and statuses

* Fixes testing dependency issue.

* Fixes Multiple Issues With Character Switching/Loading (#1340)

* Fixes issue that would cause the wizard to load into the wrong step.

* Fixes an issue where the character picker would be incorrectly shown.

* Fixes issue where a default character would be chosen.

- Default characters should no longer exist.

* Fixes multiple issues with chat. (#1342)

- Cleans up code to use more consistent conventions.
- Chat detail view is now aware of party events and updates itself accordingly.
- Chat detail view now displays a template image if no pCard is present for
  user.
- Removed some debug logging.

* Fixed issue with subscriptions.

+++ charactersheet/charactersheet/services/common/account/authentication_service.js
@@ -16,7 +16,7 @@ var AuthenticationServiceManager = new SharedServiceManager(_AuthenticationServi
 function _AuthenticationService(config) {
     var self = this;
 
-    self.validationUrl = '/api/o/validate/?access_token={access_token}';
+    self.validationUrl = '/api/o/validate/';
 
     /**
      * A configuration object that can be used to set options at initialization-time.
@@ -24,41 +24,70 @@ function _AuthenticationService(config) {
      */
     self.config = config;
 
+    self._tokenOrigin = null;
+
+    self.TOKEN_ORIGINS = {
+        FRAGMENT: 'fragment',
+        LOCAL: 'local'
+    };
+
     self.init = function() {
         // Make sure the token we have is valid.
         var fragments = (new URI()).fragment(true);
         var token = PersistenceService.findAll(AuthenticationToken)[0];
-        var url = null;
 
+        var accessToken = null;
         if (fragments['access_token']) {
-            url = self.validationUrl.replace('{access_token}', fragments['access_token']);
+            accessToken = fragments['access_token'];
+            self._tokenOrigin = self.TOKEN_ORIGINS.FRAGMENT;
         } else if (token && token.isValid()) {
-            url = self.validationUrl.replace('{access_token}', token.accessToken());
+            accessToken = token.accessToken();
+            self._tokenOrigin = self.TOKEN_ORIGINS.LOCAL;
         } else {
             return;
         }
 
-        $.getJSON(url, self._handleValidationResponse);
+        self._doAuthCheck(accessToken);
     };
 
     // Private Methods
 
+    self._doAuthCheck = function(accessToken) {
+        Utility.oauth.getJSON(self.validationUrl, self._handleValidationResponse,
+            self._handleValidationResponse, accessToken);
+    };
+
     self._handleValidationResponse = function(data, status) {
         var token = PersistenceService.findAll(AuthenticationToken)[0];
         if (status !== 'success' || data.code !== 0) {
-            if (token) {
-                token.delete();
+            if (self._tokenOrigin == self.TOKEN_ORIGINS.FRAGMENT) {
+                // Retry with local token.
+                self._tokenOrigin = self.TOKEN_ORIGINS.LOCAL;
+                if (token) {
+                    self._doAuthCheck(token.accessToken());
+                }
+            } else {
+                // All tokens are invalid and should be removed.
+                if (token) {
+                    token.delete();
+                }
             }
-        } else {
-            var fragments = (new URI()).fragment(true);
+            return;
+        }
+
+        if (self._tokenOrigin == self.TOKEN_ORIGINS.FRAGMENT) {
             if (!token) {
                 token = new AuthenticationToken();
             }
-
+            var fragments = (new URI()).fragment(true);
             token.mapTokenKeys(fragments);
             token.startTime((new Date()).getTime());
             token.save();
+            Notifications.authentication.loggedIn.dispatch();
+        } else if (self._tokenOrigin == self.TOKEN_ORIGINS.LOCAL) {
+            Notifications.authentication.loggedIn.dispatch();
+        } else {
+            return;
         }
-        Notifications.authentication.loggedIn.dispatch();
     };
 }
